fetch next bike ausgelagert | added nextBikes zu Vue watch
